Rename Inventory folder to ServiceInventory
