+++ src/Test/Inject/TablePresenter.ts
@@ -1,6 +1,6 @@
 import { DummyActionPresenter, DummyDatePresenter } from "@/Test/Mock";
-import { AttributesPresenter } from "@/UI/Inventory/Presenters/AttributesPresenter";
-import { TablePresenter } from "@/UI/Inventory/Presenters/TablePresenter";
+import { AttributesPresenter } from "@/UI/ServiceInventory/Presenters/AttributesPresenter";
+import { TablePresenter } from "@/UI/ServiceInventory/Presenters/TablePresenter";
 
 export const tablePresenter = new TablePresenter(
   new DummyDatePresenter(),
