+++ src/Test/Mock/DummyActionPresenter.ts
@@ -1,5 +1,5 @@
 import { ReactElement } from "react";
-import { ActionPresenter } from "@/UI/Inventory/Presenters";
+import { ActionPresenter } from "@/UI/ServiceInventory/Presenters";
 
 export class DummyActionPresenter implements ActionPresenter {
   getForId(): ReactElement | null {
