+++ src/UI/ServiceInventory/Components/AttributesList.tsx
