+++ src/UI/ServiceInventory/Components/AttributesSummaryView.tsx
