+++ src/UI/ServiceInventory/Components/IdWithCopy.tsx
