+++ src/UI/ServiceInventory/ExpansionManager.test.ts
