+++ src/UI/ServiceInventory/ExpansionManager.ts
