+++ src/UI/ServiceInventory/InstanceRow.tsx
