+++ src/UI/ServiceInventory/InventoryTable.stories.tsx
