+++ src/UI/ServiceInventory/InventoryTable.test.tsx
