+++ src/UI/ServiceInventory/InventoryTable.tsx
