+++ src/UI/ServiceInventory/Presenters/AttributesPresenter.test.ts
