+++ src/UI/ServiceInventory/Presenters/AttributesPresenter.ts
