+++ src/UI/ServiceInventory/Presenters/TablePresenter.ts
