+++ src/UI/ServiceInventory/TableProvider.stories.tsx
